# docs/api/openapi-specification.yaml
openapi: 3.0.0
info:
  title: Portfolio Site REST API
  version: 2.0.0
  description: |
    Comprehensive REST API for Hugo-based portfolio site with content management,
    analytics, webhooks, and GraphQL integration.
    
    ## Features
    - Full CRUD operations for all content types
    - Advanced search and filtering
    - Real-time analytics and reporting
    - Webhook system for integrations
    - GraphQL endpoint for flexible queries
    - Multi-language content support
    - Image optimization and management
    - SEO and metadata management
    
  contact:
    name: API Support
    email: support@portfolio.com
    url: https://portfolio.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.portfolio.com/v2
    description: Production server
  - url: https://staging-api.portfolio.com/v2
    description: Staging server
  - url: http://localhost:3000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /content/blog:
    get:
      tags: [Content - Blog]
      summary: List blog posts
      description: Retrieve paginated list of blog posts with filtering options
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Language'
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: author
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Blog posts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBlogPosts'

    post:
      tags: [Content - Blog]
      summary: Create new blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlogPostRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateBlogPostMultipart'
      responses:
        '201':
          description: Blog post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /content/portfolio:
    get:
      tags: [Content - Portfolio]
      summary: List portfolio items
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Language'
        - name: category
          in: query
          schema:
            type: string
            enum: [web-development, language-learning, education, tools]
        - name: technology
          in: query
          schema:
            type: array
            items:
              type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Portfolio items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPortfolioItems'

    post:
      tags: [Content - Portfolio]
      summary: Create portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortfolioRequest'
      responses:
        '201':
          description: Portfolio item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItem'
...
  /search:
    get:
      tags: [Search]
      summary: Global content search
      description: Advanced search across all content types with full-text search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 3
          description: Search query
        - name: type
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [blog, portfolio, tools, teaching, pages]
          description: Content types to search
        - name: language
          in: query
          schema:
            type: string
            enum: [en, es]
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, date, title, views]
            default: relevance
        - name: dateRange
          in: query
          schema:
            type: string
            enum: [week, month, year, all]
            default: all
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    Language:
      name: lang
      in: query
      schema:
        type: string
        enum: [en, es]
        default: en
      description: Content language
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, editor, viewer]
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
          format: uri
        isActive:
          type: boolean
        preferences:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BlogPost:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        language:
          type: string
          enum: [en, es]
        author:
          $ref: '#/components/schemas/User'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        featuredImage:
          type: string
          format: uri
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        seo:
          $ref: '#/components/schemas/SEOMetadata'
        metrics:
          $ref: '#/components/schemas/ContentMetrics'
    CreateBlogPostRequest:
      type: object
      required: [title, content, status]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        excerpt:
          type: string
          maxLength: 500
        content:
          type: string
          minLength: 1
        status:
          type: string
          enum: [draft, published]
        language:
          type: string
          enum: [en, es]
          default: en
        categories:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        featuredImage:
          type: string
          format: uri
        publishedAt:
          type: string
          format: date-time
        seo:
          $ref: '#/components/schemas/SEOMetadataInput'
    CreateBlogPostMultipart:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/CreateBlogPostRequest'
        featuredImage:
          type: string
          format: binary
        attachments:
          type: array
          items:
            type: string
            format: binary
...
    PortfolioItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [web-development, language-learning, education, tools]
        technologies:
          type: array
          items:
            type: string
        featured:
          type: boolean
        images:
          type: array
          items:
            type: string
            format: uri
        liveUrl:
          type: string
          format: uri
        repositoryUrl:
          type: string
          format: uri
        demoUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [active, completed, archived]
        language:
          type: string
          enum: [en, es]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        seo:
          $ref: '#/components/schemas/SEOMetadata'
    CreatePortfolioRequest:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        description:
          type: string
          minLength: 1
          maxLength: 500
        content:
          type: string
        category:
          type: string
          enum: [web-development, language-learning, education, tools]
        technologies:
          type: array
          items:
            type: string
        featured:
          type: boolean
          default: false
        liveUrl:
          type: string
          format: uri
        repositoryUrl:
          type: string
          format: uri
        demoUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [active, completed, archived]
          default: active
        language:
          type: string
          enum: [en, es]
          default: en
        seo:
          $ref: '#/components/schemas/SEOMetadataInput'
...
    PaginatedBlogPosts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BlogPost'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginatedPortfolioItems:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioItem'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        color:
          type: string
        itemCount:
          type: integer
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        color:
          type: string
        itemCount:
          type: integer
    SEOMetadata:
      type: object
      properties:
        metaTitle:
          type: string
          maxLength: 60
        metaDescription:
          type: string
          maxLength: 160
        metaKeywords:
          type: array
          items:
            type: string
        openGraph:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
            type:
              type: string
              default: article
        twitter:
          type: object
          properties:
            card:
              type: string
              default: summary_large_image
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
        structuredData:
          type: object
    SEOMetadataInput:
      type: object
      properties:
        metaTitle:
          type: string
          maxLength: 60
        metaDescription:
          type: string
          maxLength: 160
        metaKeywords:
          type: array
          items:
            type: string
        openGraph:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
        twitter:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
            image:
              type: string
              format: uri
    SearchResults:
      type: object
      properties:
        query:
          type: string
        total:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        facets:
          type: object
          properties:
            types:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
            categories:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
            tags:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    SearchResult:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [blog, portfolio, tools, teaching, pages]
        title:
          type: string
        excerpt:
          type: string
        url:
          type: string
        language:
          type: string
        publishedAt:
          type: string
          format: date-time
        relevanceScore:
          type: number
        highlights:
          type: array
          items:
            type: string
...
    ContentMetrics:
      type: object
      properties:
        views:
          type: integer
        uniqueViews:
          type: integer
        averageTimeOnPage:
          type: number
        bounceRate:
          type: number
        socialShares:
          type: object
          properties:
            total:
              type: integer
            twitter:
              type: integer
            facebook:
              type: integer
            linkedin:
              type: integer
        engagement:
          type: object
          properties:
            likes:
              type: integer
            comments:
              type: integer
            bookmarks:
              type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
  responses:
...
    ValidationError:
      description: Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Validation failed'
            code: 'VALIDATION_ERROR'
            details:
              email: 'Invalid email format'
              password: 'Password must be at least 8 characters'
            timestamp: '2025-01-15T10:30:00Z'
...